Moved in-file calls to projectAPI
Moved in-file calls to project API, specifically "Delete Project", used in Editor Toolbar, "Crop image" and "Delete all markers", used during cropping
