+++ frontend/components/component/CropImage.tsx
@@ -3,7 +3,7 @@ import 'react-image-crop/dist/ReactCrop.css';
 import { useState } from 'react';
 import { Button } from "@/components/ui/button";
 import CropModal from "@/components/ui/CropModal";
-import axios from 'axios';
+import * as api from "@/components/component/projectAPI";
 
 type CropImageProps = {
     onCrop: () => void;
@@ -72,52 +72,26 @@ export default function CropImage({ onCrop, resetMarkerRequest, placedMarkerAmou
             const p2x = Math.round((crop.x + crop.width) * scaleFactorX);
             const p2y = Math.round((crop.y + crop.height) * scaleFactorY);
 
-            // Fetch the Blob from the blob URL
-            const response = await fetch(imageSrc);
-            const blob = await response.blob();
+            
+            api.cropImage(imageSrc, p1x, p1y, p2x, p2y)
+                .then(() => {
+                    // Tell the current component and parent component that the image has been cropped
+                    setImageSrc(localStorage.getItem("pdfData")!);
+                    onCrop();
 
-            // Create a FormData and add the blob file
-            const formData = new FormData();
-            formData.append('file', blob, 'filename');
-
-            // Base URL for the backend API from .env
-            const BASE_URL = process.env.NEXT_PUBLIC_BACKEND_URL;
-
-            //Try to send a request to the API
-            try {
-                // Send a POST request to the API with the coordinates and file
-                const response = await axios.post(`${BASE_URL}/converter/cropPng?p1x=${p1x}&p1y=${p1y}&p2x=${p2x}&p2y=${p2y}`, formData, {
-                    headers: {
-                        'Content-Type': 'multipart/form-data'
-                    },
-                    responseType: 'blob' // Tell axios to expect a Blob in the response
+                    // Reset the crop state, use max width and height
+                    setCrop({
+                        unit: '%',
+                        x: 0,
+                        y: 0,
+                        width: 100,
+                        height: 100,
+                    });
+                })
+                .catch((error) => {
+                    // handle error
+                    console.error("Error:", error.message);
                 });
-
-                // Convert the response data to a Blob
-                const newBlob = new Blob([response.data], { type: 'image/png' });
-
-                // Convert the new Blob to a Blob URL
-                const blobUrl = URL.createObjectURL(newBlob);
-
-                // Update the source URL of the image
-                window.localStorage.setItem("pdfData", blobUrl);
-                
-                // Tell the current component and parent component that the image has been cropped
-                setImageSrc(localStorage.getItem("pdfData")!);
-                onCrop();
-
-                // Reset the crop state, use max width and height
-                setCrop({
-                    unit: '%',
-                    x: 0,
-                    y: 0,
-                    width: 100,
-                    height: 100,
-                });
-            } catch (error) {
-                // Log any errors
-                console.error('Error:', error);
-            }
 
             // Reset the apply button text
             setApplyButtonText('Apply Crop');
@@ -141,25 +115,17 @@ export default function CropImage({ onCrop, resetMarkerRequest, placedMarkerAmou
 
     // Deletes all markers from project and applies the crop
     const deleteMarkersAndApplyCrop = () => {
-        removeMarkersApiRequest(); // local function
+        console.log("Deleting project points...");
+        api.deleteAllMarkers(projectId)
+            .then(() => {
+                console.log("All markers deleted");
+            })
+            .catch((error) => {
+                console.error("Error:", error.message);
+            });   
+
         resetMarkerRequest(); // passed as prop
         handleApplyCrop();
-    };
-
-    // Contacts API and deletes all markers from project
-    const removeMarkersApiRequest = () => {
-        console.log("Deleting project points...");
-        axios.delete(`${BASE_URL}/project/${projectId}/points`, {
-            headers: {
-                'accept': 'application/json'
-            }
-        })
-        .then(response => {
-            console.log(response.data);
-        })
-        .catch(error => {
-            console.error('Error deleting data: ', error);
-        });
     };
 
     // Shows the crop modal if there are placed markers, otherwise just applies the crop
